requirements.txt
mysql-connector-python
tkcalendar
tk
pandas
numpy
matplotlib